Pagebreak -- Simple doublespaced replace
Replaced 3 carriage returns with 2. Do not want double spaced articles.

Allows more than double space, for instance between paragraphs. If this is an
issue, can explode into lines, and check in a loop. If blank line detected, turn
on a flag, and start skipping iterations with continue if further blank lines
detected