Testable -- WpConvert Unit Test, Back to Purely Functional
State in react is more than just an object property. It is not readily accessible, and more concerned
with updating DOM.

Switched backed to purely static functions, input and output only, no reading object state.

Instead of trying to chain methods, or worry about nested functions, better to focus on not
having one function depend on another and let it run async. For instance, combine pagebreaks
and bottom and top of page shortcodes for one regex, then title format, the top and bottom
article shortcodes.

This also makes it easy to factor the regex out into a utility function if needed.

Only other real option is to start adding a lot of properties to the React component,
or create an object which returns functions and holds state of it's own