Format Fixes -- Varying inner page paragraph count
Detected titles for page break and shortcode insertion.

Removed paragraph detection for lines of 70 characters or greater. Inner page paragraph count can change.
Easiest way to adjust for that is to look for the titles of 70 character or less, and use those for placement.

Regex gets complicated so that it can ignore shortcodes, detect the first and last pages when adding pagebreaks,
and ignore new lines when inserting new text